adicionado serviço de funcionario publico

+++ api/src/domain/model/public_agent.go
@@ -8,7 +8,7 @@ type PublicAgent struct {
 	Name       string    `json:"name,omitempty" xorm:"varchar(100) notnull unique(identifier)"`
 	Occupation string    `json:"occupation,omitempty" xorm:"varchar(160) notnull unique(identifier)"`
 	Department string    `json:"department,omitempty" xorm:"varchar(160) notnull unique(identifier)"`
-	Salary     float32   `json:"salary,omitempty"`
+	Salary     float64   `json:"salary,omitempty"`
 	Checked    time.Time `json:"-" xorm:"notnull"`
 	CreatedAt  time.Time `json:"-" xorm:"notnull"`
 	UpdatedAt  time.Time `json:"-" xorm:"notnull"`
